moved reusable enums and classes to core module
