+++ core/src/main/java/com/yubico/yubikit/core/UsbInterface.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Yubico.
+ * Copyright (C) 2022 Yubico.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.yubico.yubikit.management;
+package com.yubico.yubikit.core;
 
 /**
  * Provides constants for the different YubiKey USB interfaces, and the Mode enum for combinations of enabled interfaces.
